requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cloudinary==1.25.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.1.0
psycopg2==2.8.6
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.4
whitenoise==5.2.0